tools-cpi: improving default location to store results
The default location for ocount runs is being set at /home/<user>/cpi_results.
This directory will store the result files as well as the parsed output, which
is, in fact, a dictionary with 'key:value' pair being 'event_name:result'.

Also, we are adding a timestamp at the file for future comparison between runs.

Change-Id: Ib767d52b6b27dd62c6727d6a480c0658ca685846
Signed-off-by: Daniel Kreling <dbkreling@br.ibm.com>