tools-cpi: improving table readability
In this patch table_creator is changed so that the negative percentages are
shown in red, in order to highlight the points of improvement of the
application. A little formatting took place as well.

The output header was improved to provide the user how to better interpret
the table data.

Change-Id: I6d0d26d0e15e64589cc23a3dc9df0dbd19602daa
Signed-off-by: Daniel Kreling <dbkreling@br.ibm.com>