tools-cpi: changing dash character ("-") to "n/a" and "0.00"
When a percentage calculation is not possible, controller.compare_output()
draws a dash character in the screen, producing potential confusing
interpretation of the results.

This patch changes that print to "n/a" when the calculation is not possible
(e.g: division by zero). Also, when the values between runs are the same,
we print "0,00".

Change-Id: Iedda9fea5e35458bf3d3ea1a4571b707840e4ed3
Signed-off-by: Daniel Kreling <dbkreling@br.ibm.com>